+++ src/main/java/mcjty/rftools/blocks/generator/CoalGeneratorTileEntity.java
@@ -216,13 +216,13 @@ public class CoalGeneratorTileEntity extends GenericEnergyProviderTileEntity imp
 
                 if (RFTools.redstoneflux && RedstoneFluxCompatibility.isEnergyConnection(te)) {
                     if (RedstoneFluxCompatibility.canConnectEnergy(te, opposite)) {
-                        received = EnergyTools.receiveEnergy(te, opposite, rfToGive);
+                        received = (int) EnergyTools.receiveEnergy(te, opposite, rfToGive);
                     } else {
                         received = 0;
                     }
                 } else {
                     // Forge unit
-                    received = EnergyTools.receiveEnergy(te, opposite, rfToGive);
+                    received = (int) EnergyTools.receiveEnergy(te, opposite, rfToGive);
                 }
                 energyStored -= storage.extractEnergy(received, false);
                 if (energyStored <= 0) {
