+++ src/main/java/mcjty/rftools/blocks/powercell/PowerCellTileEntity.java
@@ -346,13 +346,13 @@ public class PowerCellTileEntity extends GenericTileEntity implements IEnergyPro
 
                         if (RFTools.redstoneflux && RedstoneFluxCompatibility.isEnergyConnection(te)) {
                             if (RedstoneFluxCompatibility.canConnectEnergy(te, opposite)) {
-                                received = EnergyTools.receiveEnergy(te, opposite, rfToGive);
+                                received = (int) EnergyTools.receiveEnergy(te, opposite, rfToGive);
                             } else {
                                 received = 0;
                             }
                         } else {
                             // Forge unit
-                            received = EnergyTools.receiveEnergy(te, opposite, rfToGive);
+                            received = (int) EnergyTools.receiveEnergy(te, opposite, rfToGive);
                         }
 
                         energyStored -= extractEnergyInternal(received, false, Integer.MAX_VALUE);
